create dirs for each functionality
