+++ inc/ast.h
@@ -7,7 +7,6 @@ typedef struct s_ast
 	char			*file;
 	struct s_ast	*next_left;
 	struct s_ast	*next_right;
-	// struct s_ast	*previous;
 }	t_ast;
 
 typedef enum e_order
